chore: replace table creator/dropper with Manipulator
