+++ interpreters/src/drop.rs
@@ -3,48 +3,21 @@
 //! Interpreter for drop statements
 
 use async_trait::async_trait;
-use snafu::{Backtrace, ResultExt, Snafu};
+use snafu::{ResultExt, Snafu};
 use sql::plan::DropTablePlan;
 use table_engine::engine::TableEngineRef;
 
 use crate::{
     context::Context,
     interpreter::{Drop, Interpreter, InterpreterPtr, Output, Result as InterpreterResult},
-    table_dropper::TableDropperRef,
+    table_manipulator::{self, TableManipulatorRef},
 };
 
 #[derive(Debug, Snafu)]
 #[snafu(visibility(pub(crate)))]
 pub enum Error {
-    #[snafu(display("Failed to find catalog, name:{}, err:{}", name, source))]
-    FindCatalog {
-        name: String,
-        source: catalog::manager::Error,
-    },
-
-    #[snafu(display("Catalog not exists, name:{}.\nBacktrace:\n{}", name, backtrace))]
-    CatalogNotExists { name: String, backtrace: Backtrace },
-
-    #[snafu(display("Failed to find schema, name:{}, err:{}", name, source))]
-    FindSchema {
-        name: String,
-        source: catalog::Error,
-    },
-
-    #[snafu(display("Schema not exists, name:{}.\nBacktrace:\n{}", name, backtrace))]
-    SchemaNotExists { name: String, backtrace: Backtrace },
-
-    #[snafu(display("Failed to drop table in schema, name:{}, err:{}", table, source))]
-    SchemaDropTable {
-        table: String,
-        source: catalog::schema::Error,
-    },
-
-    #[snafu(display("Failed to drop table, name:{}, err:{}", table, source))]
-    DropTable {
-        table: String,
-        source: table_engine::engine::Error,
-    },
+    #[snafu(display("Failed to drop table by table manipulator, err:{}", source))]
+    ManipulateTable { source: table_manipulator::Error },
 }
 
 define_result!(Error);
@@ -54,7 +27,7 @@ pub struct DropInterpreter {
     ctx: Context,
     plan: DropTablePlan,
     table_engine: TableEngineRef,
-    table_dropper: TableDropperRef,
+    table_manipulator: TableManipulatorRef,
 }
 
 impl DropInterpreter {
@@ -62,22 +35,23 @@ impl DropInterpreter {
         ctx: Context,
         plan: DropTablePlan,
         table_engine: TableEngineRef,
-        table_dropper: TableDropperRef,
+        table_manipulator: TableManipulatorRef,
     ) -> InterpreterPtr {
         Box::new(Self {
             ctx,
             plan,
             table_engine,
-            table_dropper,
+            table_manipulator,
         })
     }
 }
 
 impl DropInterpreter {
     async fn execute_drop(self: Box<Self>) -> Result<Output> {
-        self.table_dropper
+        self.table_manipulator
             .drop_table(self.ctx, self.plan, self.table_engine)
             .await
+            .context(ManipulateTable)
     }
 }
