+++ interpreters/src/tests.rs
@@ -19,8 +19,7 @@ use crate::{
     context::Context,
     factory::Factory,
     interpreter::{Output, Result},
-    table_creator::catalog_based::TableCreatorImpl,
-    table_dropper::catalog_based::TableDropperImpl,
+    table_manipulator::catalog_based::TableManipulatorImpl,
 };
 
 async fn build_catalog_manager(analytic: TableEngineRef) -> TableBasedManager {
@@ -60,14 +59,12 @@ where
 {
     async fn build_factory(&self) -> Factory<ExecutorImpl> {
         let catalog_manager = Arc::new(build_catalog_manager(self.engine()).await);
-        let table_creator = Arc::new(TableCreatorImpl::new(catalog_manager.clone()));
-        let table_dropper = Arc::new(TableDropperImpl::new(catalog_manager.clone()));
+        let table_manipulator = Arc::new(TableManipulatorImpl::new(catalog_manager.clone()));
         Factory::new(
             ExecutorImpl::new(query_engine::Config::default()),
             catalog_manager,
             self.engine(),
-            table_creator,
-            table_dropper,
+            table_manipulator,
         )
     }
 
@@ -139,14 +136,12 @@ where
         let ctx = Context::builder(RequestId::next_id())
             .default_catalog_and_schema(DEFAULT_CATALOG.to_string(), DEFAULT_SCHEMA.to_string())
             .build();
-        let table_creator = Arc::new(TableCreatorImpl::new(catalog_manager.clone()));
-        let table_dropper = Arc::new(TableDropperImpl::new(catalog_manager.clone()));
+        let table_manipulator = Arc::new(TableManipulatorImpl::new(catalog_manager.clone()));
         let insert_factory = Factory::new(
             ExecutorImpl::new(QueryConfig::default()),
             catalog_manager.clone(),
             self.engine(),
-            table_creator.clone(),
-            table_dropper.clone(),
+            table_manipulator.clone(),
         );
         let insert_sql = "INSERT INTO test_missing_columns_table(key1, key2, field4) VALUES('tagk', 1638428434000, 1), ('tagk2', 1638428434000, 10);";
 
@@ -165,8 +160,7 @@ where
             ExecutorImpl::new(QueryConfig::default()),
             catalog_manager,
             self.engine(),
-            table_creator,
-            table_dropper,
+            table_manipulator,
         );
         let ctx = Context::builder(RequestId::next_id())
             .default_catalog_and_schema(DEFAULT_CATALOG.to_string(), DEFAULT_SCHEMA.to_string())
