+++ server/src/instance.rs
@@ -6,7 +6,7 @@ use std::sync::Arc;
 
 use catalog::manager::ManagerRef;
 use df_operator::registry::FunctionRegistryRef;
-use interpreters::{table_creator::TableCreatorRef, table_dropper::TableDropperRef};
+use interpreters::table_manipulator::TableManipulatorRef;
 use table_engine::engine::TableEngineRef;
 
 use crate::limiter::Limiter;
@@ -21,8 +21,7 @@ pub struct Instance<Q> {
     // User defined functions registry.
     pub function_registry: FunctionRegistryRef,
     pub limiter: Limiter,
-    pub table_creator: TableCreatorRef,
-    pub table_dropper: TableDropperRef,
+    pub table_manipulator: TableManipulatorRef,
 }
 
 /// A reference counted instance pointer
